rename : toastMessage => errorMessage로 다시 변경
